requirements.txt
sqlalchemy~=2.0.31
pydantic~=2.8.2
fastapi~=0.111.1
passlib~=1.7.4
PyJWT~=2.8.0
bcrypt~=4.2.0
alembic~=1.13.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
uvicorn==0.30.1